backend/requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
boto3==1.12.31
botocore==1.15.31
celery==4.4.2
Django==3.0.4
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
importlib-metadata==1.6.0
jmespath==0.9.5
kombu==4.6.8
python-dateutil==2.8.1
python-http-client==3.2.6
pytz==2019.3
s3transfer==0.3.3
sendgrid==6.2.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
